Add stock_picker function
This commit introduces a new function called 'stock_picker' which takes an array of stock prices as an input and returns the best days to buy and sell in order to make the maximum profit. The function traverses the array keeping track of the minimum price and the maximum profit simultaneously.